Package up some of the auto pairing stuff in list wizard for reuse.
