+++ client/src/components/Collections/ListWizard.vue
@@ -17,9 +17,8 @@ import {
 } from "./common/useCollectionCreator";
 import { showHid } from "./common/useCollectionCreator";
 import type { WhichListBuilder } from "./ListWizard/types";
-import { autoPairWithCommonFilters } from "./pairing";
+import { useAutoPairing } from "./usePairing";
 
-import AutoPairing from "./common/AutoPairing.vue";
 import ListCollectionCreator from "./ListCollectionCreator.vue";
 import WhichBuilder from "./ListWizard/WhichBuilder.vue";
 import PairedOrUnpairedListCollectionCreator from "./PairedOrUnpairedListCollectionCreator.vue";
@@ -36,8 +35,8 @@ const isBusy = ref<boolean>(false);
 const whichBuilder = ref<WhichListBuilder>("list");
 const store = useCollectionBuilderItemSelection();
 const { currentHistoryId, createPayload } = useCollectionCreation();
-const currentForwardFilter = ref("");
-const currentReverseFilter = ref("");
+const { countPaired, currentForwardFilter, currentReverseFilter, AutoPairing, autoPair, onFilters } = useAutoPairing();
+
 const creationError = ref<string | null>(null);
 const collectionCreated = ref(false);
 
@@ -45,18 +44,13 @@ type InferrableBuilder = "list" | "list:paired";
 const collectionCreator = ref<CollectionCreatorComponent>();
 const { selectedItems } = storeToRefs(store);
 
-const countPaired = ref(-1);
-const countUnpaired = ref(-1);
 const inferredBuilder = ref<InferrableBuilder>("list");
 const builderInputsValid = ref(false);
 
 async function initialize() {
     isBusy.value = true;
-    const summary = autoPairWithCommonFilters(selectedItems.value, true);
-    currentForwardFilter.value = summary.forwardFilter || "";
-    currentReverseFilter.value = summary.reverseFilter || "";
-    countPaired.value = summary.pairs?.length || 0;
-    countUnpaired.value = summary.unpaired.length;
+    autoPair(selectedItems.value);
+    // updates currentForwardFilter, currentReverseFilter, countPaired
     if (countPaired.value * 2 > selectedItems.value.length * 0.2) {
         whichBuilder.value = "list:paired";
         inferredBuilder.value = "list:paired";
@@ -189,11 +183,6 @@ async function ruleOnAttemptCreate(createRequest: RuleCreationRequestT) {
 
 function setWhichBuilder(newWhichBuilder: WhichListBuilder) {
     whichBuilder.value = newWhichBuilder;
-}
-
-function onFilters(forwardFilter: string, reverseFilter: string) {
-    currentForwardFilter.value = forwardFilter;
-    currentReverseFilter.value = reverseFilter;
 }
 
 function goToAutoPairing() {
